[wsu] Configure CNI on bootstrapped nodes
This commit run the wmcb's configure-cni command on the bootstrapped
nodes. In addition it distributes the files needed to run the command to
the nodes.

This commit also adds an e2e test to the wsu test suite, which deploys
Windows and Linux pods, and ensures they can communicate between each
other.