Bug 1787646: Update hybrid-overlay.exe for 4.4
Problem:
The execution of the hybrid-overlay.exe is failing in the Windows node
as it cannot find the required "k8s.ovn.org/hybrid-overlay-hostsubnet"
annotation on the node object that should be applied on enabling hybrid
OVN networking.

Solution:
The hybrid annotation has changed to
"k8s.ovn.org/hybrid-overlay-node-subnet" in 4.4 and to pick those
changes up we need to use a newer version of the hybrid-overlay.exe. A
release was cut that included the newer version built from the master
(4.4) branch of openshift/ovn-kubernetes and WSU was updated to use this
version.